backend/requirements.txt
requests
aiohttp
python-dotenv
livekit-server-sdk
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
